refactor: extract calimero-governance-types leaf crate to break actix cycle (#2479)
Part of epic #2300. Prereq for #2307 (the physical extraction of
`calimero-governance-store`): without this cycle break, the new
crate would re-acquire `actix` transitively through
`calimero-context-client`, defeating the whole point of the
extraction.

## What moved

The op-data half of `calimero-context-client::local_governance/`
moves out into a new leaf crate `calimero-governance-types`:

  calimero-context-client/src/local_governance/mod.rs   → calimero-governance-types/src/lib.rs
  calimero-context-client/src/local_governance/wire.rs  → calimero-governance-types/src/wire.rs
  calimero-context-client/src/local_governance/tests.rs → calimero-governance-types/src/tests.rs

Plus a new `crates/governance-types/Cargo.toml` and workspace
member entry. `git mv` used throughout so blame is preserved.

## What stays in calimero-context-client

`local_governance/ack_router.rs` — the `AckRouter` pub/sub primitive
that uses `tokio::sync::broadcast` to route signed acks between
the publish-side helper in `calimero-context` and the receive-side
gossipsub handler in `calimero-node`. It's runtime plumbing that
connects actor mailboxes, not a data type. Leaving it in
`calimero-context-client` means `calimero-governance-types` can
stay actix-free.

A new `local_governance/mod.rs` is created as a thin deprecation
shim: it `mod ack_router; pub use ack_router::AckRouter;` for the
local piece, plus `#[deprecated]` re-exports of every symbol
previously exposed from the old `mod.rs` — pointing callers at
`calimero_governance_types::*`. External consumers continue to
compile unchanged; they just see one `#[deprecated]` warning per
imported symbol nudging them to migrate.

## Acyclicity check

The whole point of this PR:

  $ cargo tree -p calimero-governance-types -e features | grep actix
  (empty — clean)

`calimero-governance-types`'s transitive deps are blake3, borsh,
ed25519-dalek, sha2, thiserror, `calimero-primitives`, and
`calimero-context-config` (which is itself actix-free). No
`actix*`, no `tokio` (the moved code didn't need it; only
`ack_router.rs` did, and that stayed). When #2307 lands and
`calimero-governance-store` depends on `calimero-governance-types`,
the read path of `crates/server` and `crates/node` (23 files
identified in the #2300 audit) can drop their `calimero-context`
dep without losing access to the op types.

## In-crate touch-ups

`ack_router.rs` previously imported `SignedAck` via `super::SignedAck`
(referring to the old mod.rs's local definition). After move,
that path resolves through the deprecation shim and would emit a
deprecation warning at the import site. Switched to a direct
`use calimero_governance_types::SignedAck;` to keep the in-crate
code clean of self-deprecation churn.

## Verification

- `cargo tree -p calimero-governance-types -e features` contains
  no `actix*`.
- `cargo build --workspace` green.
- `cargo test --workspace --lib` — every package's test count is
  unchanged. Of note: 26 tests in `calimero-governance-types` (the
  moved `tests.rs`), 315 in `calimero-context`, 213 in `calimero-node`,
  7 in `calimero-context-client`. Total workspace lib tests stay at
  the same total.
- `cargo fmt --all -- --check` clean.
- `git mv` preserved blame on the three moved files.

## Out of scope

- Migrating call sites off the deprecation shim — separate cleanup
  after one release cycle, once external consumers have absorbed
  the deprecation warnings.
- Splitting the new crate further (e.g. separating wire from ops).
  Start at one crate; subdivide only if shape demands it.
