+++ Cargo.toml
@@ -46,6 +46,7 @@ members = [
     "./crates/context",
     "./crates/context/config",
     "./crates/context/primitives",
+    "./crates/governance-types",
     "./crates/crypto",
     "./crates/dag",
     "./crates/meroctl",
@@ -225,6 +226,7 @@ calimero-config = { path = "./crates/config" }
 calimero-context = { path = "./crates/context" }
 calimero-context-config = { path = "./crates/context/config", default-features = false }
 calimero-context-client = { path = "./crates/context/primitives" }
+calimero-governance-types = { path = "./crates/governance-types" }
 calimero-crypto = { path = "./crates/crypto" }
 calimero-dag = { path = "./crates/dag" }
 calimero-network = { path = "./crates/network" }
